project/requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
bleach[css]==6.1.0
boto3==1.28.62
botocore==1.31.62; python_version >= '3.7'
django==4.2.6
django-rest-framework==0.1.0
django-storages[s3]==1.14.2
django-taggit==4.0.0
djangorestframework==3.14.0
gunicorn==21.2.0
jmespath==1.0.1; python_version >= '3.7'
markdown-it-py==3.0.0
mdit-py-plugins==0.4.0
mdurl==0.1.2; python_version >= '3.7'
packaging==23.2; python_version >= '3.7'
pillow==10.0.1
psycopg2-binary==2.9.9
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
pytz==2023.3.post1
s3transfer==0.7.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlparse==0.4.4; python_version >= '3.5'
tinycss2==1.2.1
urllib3==2.0.6; python_version >= '3.10'
webencodings==0.5.1
